refactor: simplify reset password logic
