+++ apps/api/.env.test
@@ -7,7 +7,7 @@ STRIPE_SECRET_KEY=
 REDIS_URL=
 POSTHOG_API_KEY=
 RESEND_API_KEY=
-RESEND_FROM_EMAIL=noreply@conar.app 
+RESEND_FROM_EMAIL=noreply@conar.app
 LOOPS_API_KEY=
 OPENAI_API_KEY=
 ANTHROPIC_API_KEY=
