+++ apps/api/src/orpc/routers/contact.ts
@@ -1,6 +1,6 @@
 import { SUPPORT_EMAIL } from '@conar/shared/constants'
 import { type } from 'arktype'
-import { resend } from '~/lib/email/config/resend'
+import { resend } from '~/lib/email'
 import { authMiddleware, orpc } from '~/orpc'
 
 export const contact = orpc
